move version into its own pkg
